+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning][].
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
+## [Unreleased]
+
+### New Features
+
+-   `dso watermark` now supports files in PDF format. With this change, quarto reports using the watermark feature can
+    be rendered to PDF, too.
 
 ## v0.8.2